Dekamoji is a React component that displays a very large character.
Improvement point: When displaying a long sentence, the size of the character is automatically adjusted.
Until now, the calculation was centered on Japanese full-width characters.

Also, when the sentence is long, it was not possible to make the characters larger by wrapping them.
This time, I will solve this problem.