Use kitsuyui/js-rectangle-dividing which is a spin-off project
https://github.com/kitsuyui/js-rectangle-dividing

The algorithm to divide a rectangle is a concept of pure numerical calculation rather than a React component.
Therefore, it is better to manage it in a different repository than to manage it in this repository.
I implemented it that way.
Thanks to that, the test coverage of the separation destination has also increased, and the abstraction has also progressed.