Consider font in Dekamoji
Dekamoji is a component that displays characters by enlarging them to the maximum size of the area.
Until now, Dekamoji did not consider fonts when measuring, so it was modified to consider fonts.
Get the font inherited from the parent component and use it to measure.

use getComputedStyle to get font