Implement a component for displaying binary data.
Usage: `<DetailedBitString value={0b10101100} length={16} />`

The `value` prop is the binary value to display. The `length` prop is the number of bits to display.
When `length` is not specified, the component will display the minimum number of bits to represent the value.
If actual value is longer than `length`, the component will display the overflow bits as red text.