Change the font size calculation of Dekamoji to binary search
In the previous implementation, the font size was adjusted by increasing it in 1px increments.
This caused performance degradation because the calculation took a lot of time as the font size increased.
This PR changes the font size adjustment to use binary search. This improves the calculation time from O(n) to O(log n).